Actually set XCR0
I'm pretty sure this is at least a step in the right direction, but it's
causing way more failures because KVM won't let me enable AVX in XRC0.
I think that might be because I'm on a fairly old kernel on my
laptop, but new ones won't boot and I don't want to lose this all when I
go home.

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>